flask-service/requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.84.0
pytesseract==0.3.10
pillow==10.0.0
PyPDF2==3.0.1
pdf2image==1.16.3